backend/requirements.txt
django==4.2.13
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.1
celery==5.3.6
redis==5.0.4
django-celery-beat==2.6.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.1.0
boto3==1.34.69
django-storages==1.14.2
Pillow==10.3.0
requests==2.31.0
django-encrypted-model-fields==0.6.5
django-ratelimit==4.1.0
gunicorn==21.2.0
whitenoise==6.6.0
